+++ R/ebola_kikwit_1995.R
@@ -14,7 +14,7 @@
 #'
 #' @docType data
 #'
-#' @format A data frame with 139 rows and 3 columns
+#' @format A data frame with 192 rows and 3 columns
 #' \describe{
 #'   \item{date_of_onset}{Date of onset}
 #'   \item{incidence}{Number of new cases}
@@ -37,8 +37,29 @@
 #'
 #' @family ebola
 #'
+#' @seealso \code{\link{ebola_kikwit_1995_as_ts}}
+#'
 #' @examples
 #' ## show first few cases
 #' head(ebola_kikwit_1995)
 #'
 "ebola_kikwit_1995"
+
+
+#' Coerce the dataset to a `tsibble` object
+#'
+#' Ensure it is a valid timeseries (regular, ordered,
+#' no duplication, no gaps)
+#'
+#' @return The dataset as a `tsibble` if the `tsibble` is installed, else the
+#' dataset
+#' @export
+#'
+#' @examples
+#' ebola_kikwit_1995_as_ts()
+ebola_kikwit_1995_as_ts <- function() {
+  ebola_kikwit_1995 <- outbreaks::ebola_kikwit_1995
+  if (requireNamespace("tsibble", quietly = TRUE)) {
+    tsibble::as_tsibble(ebola_kikwit_1995, index = "date_of_onset")
+  }
+}