+++ R/outbreaks-package.R
@@ -1,7 +1,7 @@
-library(tools)
 # Import from tibble in order to get consistent behaviour regardless of
 # whether or not tibble is attached.
 #' @importFrom tibble tibble
+#' @import magrittr
 NULL