+++ data-raw/ebola_kikwit.R
@@ -22,9 +22,15 @@ leg_ebola_kikwit_1995 %>% group_by(reporting) %>% summarise_if(is.numeric, sum)
 # Creation of the new dataset -----
 
 ebola_kikwit_1995 <- leg_ebola_kikwit_1995 %>%
-  filter(reporting) %>%
-  select(-reporting) %>%
   rename(date_of_onset = date, incidence = onset) %>%
+  # TODO: could be factorized in one line ...
+  mutate(incidence = case_when(reporting ~ incidence,
+                               TRUE ~ NA_integer_)) %>%
+  mutate(death = case_when(reporting ~ death,
+                        TRUE ~ NA_integer_)) %>%
+  select(-reporting) %>%
+  arrange(date_of_onset) %>%
   as_tibble()
 
 usethis::use_data(ebola_kikwit_1995, overwrite = TRUE, compress = 'xz')
+